// src/infrastructure/database/enums/providers.enums.ts
export enum ProvidersEnums {
  POST_MODEL = 'POST_MODEL',
  BLOG_MODEL = 'BLOG_MODEL',
  BBLOG_MODEL = 'BBLOG_MODEL',
  USER_MODEL = 'USER_MODEL',
  COMMENT_MODEL = 'COMMENT_MODEL',
  LIKE_STATUS_COMMENTS_MODEL = 'LIKE_STATUS_COMMENTS_MODEL',
  LIKE_STATUS_POSTS_MODEL = 'LIKE_STATUS_POSTS_MODEL',
  CONFIRM_CODE_MODEL = 'CONFIRM_CODE_MODEL',
  BL_REFRESH_JWT_MODEL = 'BL_REFRESH_JWT_MODEL',
  BL_ACCESS_JWT_MODEL = 'BL_ACCESS_JWT_MODEL',
  DEVICES_MODEL = 'DEVICES_MODEL',
}